I added more information on IOTA protocols and protocol design
I added some more information and references about MAM / IOTA
protocol. I also added a section for protocol desgin where I explain
that I was thinking about using BEEP as a basis for our protocol, but
then I discovered that BEEP lost its community in 2016 so it might not
be a worthwhile effort to try and use BEEP as a framework. I also
created individual sections for the different protocols.