Added turtle mock library for unit testing
I added the turtle mock library. It seems to be a mock library that
lots of people are very happy with. Unfortunately the maintainer has
lost interest in this library since he is no longer working on c++
projects. The maintainer is still taking PR though.

I added a pragma to explicitly ignore warnings from turtle. This
pragma only works with GCC diagnostics. I tried to ignore these
warnings from the bazel workspace using copts, but that does not
appear to be working for some reason.